requirements.txt
appdirs==1.4.4
asgiref==3.2.10
certifi==2020.11.8
chardet==3.0.4
distlib==0.3.1
Django==3.1.2
django-debug-toolbar==3.1.1
django-embed-video==1.3.3
django-extensions==3.0.9
django-ranged-response==0.2.0
django-simple-captcha==0.5.13
filelock==3.0.12
idna==2.10
Pillow==8.0.1
pytz==2020.1
requests==2.25.0
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.2
virtualenv==20.1.0
Werkzeug==1.0.1
